# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FreightFlow - Freight Invoice Automation & Reconciliation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus];

export const InvoiceStatus = {
  pending: "pending",
  matched: "matched",
  disputed: "disputed",
  paid: "paid",
  overdue: "overdue",
} as const;

export interface Invoice {
  id: number;
  invoiceNumber: string;
  vendorId: number;
  vendorName: string;
  shipmentId?: number | null;
  amount: number;
  currency: string;
  status: InvoiceStatus;
  invoiceDate: string;
  dueDate: string;
  description?: string | null;
  freightCharges: number;
  fuelSurcharge?: number | null;
  otherCharges?: number | null;
  gstAmount?: number | null;
  discrepancyNotes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface InvoiceListResponse {
  data: Invoice[];
  total: number;
  page: number;
  limit: number;
}

export interface CreateInvoiceRequest {
  invoiceNumber: string;
  vendorId: number;
  shipmentId?: number | null;
  amount: number;
  currency?: string;
  invoiceDate: string;
  dueDate: string;
  description?: string | null;
  freightCharges: number;
  fuelSurcharge?: number | null;
  otherCharges?: number | null;
  gstAmount?: number | null;
}

export type UpdateInvoiceRequestStatus =
  (typeof UpdateInvoiceRequestStatus)[keyof typeof UpdateInvoiceRequestStatus];

export const UpdateInvoiceRequestStatus = {
  pending: "pending",
  matched: "matched",
  disputed: "disputed",
  paid: "paid",
  overdue: "overdue",
} as const;

export interface UpdateInvoiceRequest {
  status?: UpdateInvoiceRequestStatus;
  shipmentId?: number | null;
  discrepancyNotes?: string | null;
  amount?: number;
  fuelSurcharge?: number | null;
  otherCharges?: number | null;
  gstAmount?: number | null;
}

export type VendorCategory =
  (typeof VendorCategory)[keyof typeof VendorCategory];

export const VendorCategory = {
  road: "road",
  rail: "rail",
  air: "air",
  sea: "sea",
  multimodal: "multimodal",
} as const;

export interface Vendor {
  id: number;
  name: string;
  gstin?: string | null;
  contactEmail?: string | null;
  contactPhone?: string | null;
  address?: string | null;
  paymentTermsDays: number;
  category: VendorCategory;
  totalInvoices?: number | null;
  totalAmountDue?: number | null;
  createdAt: string;
}

export type CreateVendorRequestCategory =
  (typeof CreateVendorRequestCategory)[keyof typeof CreateVendorRequestCategory];

export const CreateVendorRequestCategory = {
  road: "road",
  rail: "rail",
  air: "air",
  sea: "sea",
  multimodal: "multimodal",
} as const;

export interface CreateVendorRequest {
  name: string;
  gstin?: string | null;
  contactEmail?: string | null;
  contactPhone?: string | null;
  address?: string | null;
  paymentTermsDays?: number;
  category: CreateVendorRequestCategory;
}

export interface UpdateVendorRequest {
  name?: string;
  gstin?: string | null;
  contactEmail?: string | null;
  contactPhone?: string | null;
  address?: string | null;
  paymentTermsDays?: number;
}

export type ShipmentStatus =
  (typeof ShipmentStatus)[keyof typeof ShipmentStatus];

export const ShipmentStatus = {
  in_transit: "in_transit",
  delivered: "delivered",
  delayed: "delayed",
  cancelled: "cancelled",
} as const;

export interface Shipment {
  id: number;
  shipmentNumber: string;
  vendorId: number;
  vendorName: string;
  origin: string;
  destination: string;
  status: ShipmentStatus;
  scheduledDelivery?: string | null;
  actualDelivery?: string | null;
  agreedFreightCost: number;
  weightKg?: number | null;
  description?: string | null;
  createdAt: string;
}

export interface CreateShipmentRequest {
  shipmentNumber: string;
  vendorId: number;
  origin: string;
  destination: string;
  scheduledDelivery?: string | null;
  agreedFreightCost: number;
  weightKg?: number | null;
  description?: string | null;
}

export type UpdateShipmentRequestStatus =
  (typeof UpdateShipmentRequestStatus)[keyof typeof UpdateShipmentRequestStatus];

export const UpdateShipmentRequestStatus = {
  in_transit: "in_transit",
  delivered: "delivered",
  delayed: "delayed",
  cancelled: "cancelled",
} as const;

export interface UpdateShipmentRequest {
  status?: UpdateShipmentRequestStatus;
  actualDelivery?: string | null;
  agreedFreightCost?: number;
}

export interface ReconciliationRequest {
  invoiceIds: number[];
}

export interface ReconciliationDetail {
  invoiceId: number;
  invoiceNumber: string;
  status: string;
  varianceAmount: number;
  reason: string;
}

export interface ReconciliationResult {
  processed: number;
  matched: number;
  disputed: number;
  discrepanciesFound: number;
  totalVarianceAmount: number;
  details: ReconciliationDetail[];
}

export type DiscrepancyType =
  (typeof DiscrepancyType)[keyof typeof DiscrepancyType];

export const DiscrepancyType = {
  amount_mismatch: "amount_mismatch",
  duplicate: "duplicate",
  missing_shipment: "missing_shipment",
  rate_deviation: "rate_deviation",
  extra_charges: "extra_charges",
} as const;

export type DiscrepancyStatus =
  (typeof DiscrepancyStatus)[keyof typeof DiscrepancyStatus];

export const DiscrepancyStatus = {
  open: "open",
  resolved: "resolved",
  escalated: "escalated",
} as const;

export interface Discrepancy {
  id: number;
  invoiceId: number;
  invoiceNumber: string;
  vendorName: string;
  type: DiscrepancyType;
  invoicedAmount: number;
  expectedAmount: number;
  varianceAmount: number;
  status: DiscrepancyStatus;
  description?: string | null;
  resolution?: string | null;
  createdAt: string;
}

export type ResolveDiscrepancyRequestStatus =
  (typeof ResolveDiscrepancyRequestStatus)[keyof typeof ResolveDiscrepancyRequestStatus];

export const ResolveDiscrepancyRequestStatus = {
  resolved: "resolved",
  escalated: "escalated",
} as const;

export interface ResolveDiscrepancyRequest {
  status: ResolveDiscrepancyRequestStatus;
  resolution: string;
}

export interface DashboardStats {
  totalInvoices: number;
  pendingInvoices: number;
  disputedInvoices: number;
  overdueInvoices: number;
  totalAmountPending: number;
  totalAmountDisputed: number;
  totalAmountOverdue: number;
  reconciliationRate: number;
  openDiscrepancies: number;
  totalVendors: number;
  activeShipments: number;
  delayedShipments: number;
}

export interface TrendDataPoint {
  period: string;
  totalFreightCost: number;
  invoiceCount: number;
  avgInvoiceAmount: number;
  disputeRate: number;
}

export type ListInvoicesParams = {
  status?: ListInvoicesStatus;
  vendorId?: number;
  page?: number;
  limit?: number;
};

export type ListInvoicesStatus =
  (typeof ListInvoicesStatus)[keyof typeof ListInvoicesStatus];

export const ListInvoicesStatus = {
  pending: "pending",
  matched: "matched",
  disputed: "disputed",
  paid: "paid",
  overdue: "overdue",
} as const;

export type ListShipmentsParams = {
  status?: ListShipmentsStatus;
  vendorId?: number;
};

export type ListShipmentsStatus =
  (typeof ListShipmentsStatus)[keyof typeof ListShipmentsStatus];

export const ListShipmentsStatus = {
  in_transit: "in_transit",
  delivered: "delivered",
  delayed: "delayed",
  cancelled: "cancelled",
} as const;

export type ListDiscrepanciesParams = {
  status?: ListDiscrepanciesStatus;
};

export type ListDiscrepanciesStatus =
  (typeof ListDiscrepanciesStatus)[keyof typeof ListDiscrepanciesStatus];

export const ListDiscrepanciesStatus = {
  open: "open",
  resolved: "resolved",
  escalated: "escalated",
} as const;

export type GetFreightTrendsParams = {
  period?: GetFreightTrendsPeriod;
};

export type GetFreightTrendsPeriod =
  (typeof GetFreightTrendsPeriod)[keyof typeof GetFreightTrendsPeriod];

export const GetFreightTrendsPeriod = {
  week: "week",
  month: "month",
  quarter: "quarter",
} as const;