# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FreightFlow - Freight Invoice Automation & Reconciliation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List all vendors/carriers
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a new vendor/carrier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorRequest"
      responses:
        "201":
          description: Vendor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVendorRequest"
      responses:
        "200":
          description: Vendor updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
...
components:
  schemas:
...
    Vendor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        gstin:
          type: string
          nullable: true
        contactEmail:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        paymentTermsDays:
          type: integer
          default: 30
        category:
          type: string
          enum: [road, rail, air, sea, multimodal]
        totalInvoices:
          type: integer
          nullable: true
        totalAmountDue:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, paymentTermsDays, category, createdAt]
    CreateVendorRequest:
      type: object
      properties:
        name:
          type: string
        gstin:
          type: string
          nullable: true
        contactEmail:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        paymentTermsDays:
          type: integer
          default: 30
        category:
          type: string
          enum: [road, rail, air, sea, multimodal]
      required: [name, category]
    UpdateVendorRequest:
      type: object
      properties:
        name:
          type: string
        gstin:
          type: string
          nullable: true
        contactEmail:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        paymentTermsDays:
          type: integer
...