# lib/api-zod/src/generated/types/vendor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FreightFlow - Freight Invoice Automation & Reconciliation API
 * OpenAPI spec version: 0.1.0
 */
import type { VendorCategory } from "./vendorCategory";

export interface Vendor {
  id: number;
  name: string;
  gstin?: string | null;
  contactEmail?: string | null;
  contactPhone?: string | null;
  address?: string | null;
  paymentTermsDays: number;
  category: VendorCategory;
  totalInvoices?: number | null;
  totalAmountDue?: number | null;
  createdAt: Date;
}